Ticket #1252
Frontend Unit Test Cases (Improve Test Coverage to 80%)
