+++ src/components/ComponentPages/ManageSupport/ManageSupportUI/__test__/manageSupport.test.tsx
@@ -5,12 +5,7 @@ import {
   waitFor,
 } from "../../../../../utils/testUtils";
 import messages from "../../../../../messages";
-import ManageSupportUI from "../index";
-import { CloseCircleOutlined } from "@ant-design/icons";
 import ManageSupport from "../..";
-import { Card } from "antd";
-import userEvent from "@testing-library/user-event";
-import { placeholders } from "src/messages/placeholder";
 
 const { labels } = messages;
 const nickNameList = [
@@ -119,27 +114,37 @@ jest.mock("react-redux", () => ({
       asofdate: "",
       manageSupportUrlLink: "",
       currentDelegatedSupportedClick: {
-        delegatedSupportClick: false,
+        delegatedSupportClick: true,
       },
       currentGetCheckSupportExistsData: {
-        warning: "",
+        warning: "ABC",
         is_confirm: 1,
-        remove_camps: [
-          {
-            camp: "abc",
-          },
-        ],
+        remove_camps: [{camp_name:"ABC",camp_num:1,link:"/topic",support_order:1}],
+        support_flag: 1,
+        is_delegator: 1,
       },
-      CurrentCheckSupportStatus: "",
+      CurrentCheckSupportStatus: "ABC",
       manageSupportStatusCheck: true,
     };
   }),
 }));
+
 jest.mock("src/hooks/isUserAuthenticated", () =>
   jest.fn(() => ({ isUserAuthenticated: true }))
 );
 jest.mock("src/network/api/campDetailApi", () => ({
-  getAllUsedNickNames: jest.fn(),
+  getAllUsedNickNames: jest
+  .fn()
+  .mockReturnValue(Promise.resolve({
+    status_code: 200,
+    data: [{
+      create_time: "1970-01-01",
+      id: "1",
+      nick_name: "ABC",
+      owner_code: "qwerty",
+      private: 0,
+    }]  ,
+  })),
   getCurrentCampRecordApi: jest
     .fn()
     .mockReturnValue(Promise.resolve({ camp_name: "Test Camp name" })),
@@ -152,14 +157,18 @@ jest.mock("src/network/api/campDetailApi", () => ({
 }));
 jest.mock("src/network/api/topicAPI", () => ({
   GetActiveSupportTopic: jest.fn(() =>
-    Promise.resolve({ data: [], status_code: 200 })
+    Promise.resolve({ data: manageSupportList, status_code: 200 })
   ),
   GetCheckSupportExists: jest.fn(() =>
     Promise.resolve({
-      data: {
-        remove_camps: [],
-      },
       status_code: 200,
+      // manageSupportStatusCheck:false,
+      data: {
+        warning: "ABC",
+        support_flag: 0,
+        remove_camp:[{camp_name:"ABC",camp_num:1,link:"/topic",support_order:1}],
+        disable_submit:true
+      },
     })
   ),
 }));
@@ -173,6 +182,7 @@ jest.mock("src/network/api/userApi", () => ({
       status_code: 200,
     })
   ),
+  addDelegateSupportCamps: jest.fn(() => Promise.resolve({ status_code: 200, data: [{}] })),
 }));
 const supportRemovedModal = () => <div>Removed Modal</div>;
 supportRemovedModal.displayName = "SupportRemovedModal";
@@ -184,468 +194,6 @@ jest.mock(
   { virtual: true }
 );
 
-describe("ManageSupportUI", () => {
-  it("render show SupportedCamps", () => {
-    render(
-      <ManageSupportUI
-        nickNameList={nickNameList}
-        manageSupportList={manageSupportList}
-        clearAllChanges={clearAllChanges}
-        removeAll={removeAll}
-        handleClose={handleClose}
-        checked={checked}
-        setManageSupportList={setManageSupportList}
-        parentSupportDataList={parentSupportDataList}
-        getSupportStatusData={getSupportStatusData}
-        submitNickNameSupportCamps={submitNickNameSupportCamps}
-        cancelManageRoute={cancelManageRoute}
-        setSelectedtNickname={setSelectedtNickname}
-        selectedtNickname={selectedtNickname}
-        submitButtonDisable={submitButtonDisable}
-        setUpdatePostion={setUpdatePostion}
-        unableToFindCamp={unableToFindCamp}
-      />
-    );
-    expect(screen.getByText(labels.SupportedCamps)).toBeTruthy();
-  });
-});
-
-describe("ManageSupportUI Nick name", () => {
-  it("render show Nickname To Support Above Camps", () => {
-    render(
-      <ManageSupportUI
-        nickNameList={nickNameList}
-        manageSupportList={manageSupportList}
-        clearAllChanges={clearAllChanges}
-        removeAll={removeAll}
-        handleClose={handleClose}
-        checked={checked}
-        setManageSupportList={setManageSupportList}
-        parentSupportDataList={parentSupportDataList}
-        getSupportStatusData={getSupportStatusData}
-        submitNickNameSupportCamps={submitNickNameSupportCamps}
-        cancelManageRoute={cancelManageRoute}
-        setSelectedtNickname={setSelectedtNickname}
-        selectedtNickname={selectedtNickname}
-        submitButtonDisable={submitButtonDisable}
-        setUpdatePostion={setUpdatePostion}
-        unableToFindCamp={unableToFindCamp}
-      />
-    );
-    expect(screen.getByText("Nickname To Support Above Camps")).toBeTruthy();
-  });
-});
-
-it("render show submit button", () => {
-  const { getAllByText } = render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  const submitButton = getAllByText("Submit")[0] as HTMLButtonElement;
-  expect(submitButton).toBeTruthy();
-});
-
-it("render show cancel button", () => {
-  const { getAllByText } = render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  const cancelButton = getAllByText("Cancel")[0] as HTMLButtonElement;
-  expect(cancelButton).toBeTruthy();
-});
-/* eslint-disable */
-
-it("render support remove card component", () => {
-  const manageSupportList = [
-    {
-      camp_name: "Test Camp name",
-      camp_num: 2,
-      id: 1,
-      link: "/topic/949-top/2-camp-1",
-      support_order: 1,
-      topic_num: 10,
-    },
-  ];
-  const mockRemoveAll = jest.fn();
-  render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={mockRemoveAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  // eslint-disable-next-line
-  const checkboxElement = screen.getByTestId("checkbox");
-  fireEvent.click(checkboxElement);
-});
-/* eslint-enable */
-
-it("render show clear changes button", () => {
-  const { getAllByText } = render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  const clearChangesButton = getAllByText(
-    "Clear all changes"
-  )[0] as HTMLButtonElement;
-  expect(clearChangesButton).toBeTruthy();
-  expect(<CloseCircleOutlined />).toBeTruthy();
-});
-
-it("render show Quick Action Text", () => {
-  render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  expect(screen.getByText("Quick Actions:")).toBeTruthy();
-  expect(messages).toBeDefined();
-});
-
-it("render show Remove all Text", () => {
-  const { getAllByText } = render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  const removeAllText = getAllByText("Remove all");
-  expect(removeAllText).toBeTruthy();
-});
-
-it("render show checkbox", () => {
-  const { container } = render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  expect(
-    container.getElementsByClassName("ManageSupport_checkbox__DQcrk")
-  ).toBeTruthy();
-});
-
-it("render show the text line when the topic list array is not empty", () => {
-  render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  expect(screen.queryByTestId(labels.topicSupportText)).toBeNull();
-});
-
-it("render show manage support note", () => {
-  render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  expect(screen.getByText(labels.manageSupportNote)).toBeTruthy();
-});
-
-it("render show dropdown", () => {
-  const { container } = render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  expect(
-    container.getElementsByClassName("ant-select-selection-item")
-  ).toBeTruthy();
-});
-
-it("Check the checkbox is checked/unchecked", () => {
-  const { container } = render(
-    <ManageSupportUI
-      nickNameList={nickNameList}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-    />
-  );
-  expect(
-    container.getElementsByClassName("ant-select-selection-item")
-  ).toBeTruthy();
-  expect(
-    container.getElementsByClassName("ManageSupport_checkbox__DQcrk")
-  ).toBeTruthy();
-});
-
-// it("should render card with title and content", () => {
-//   render(
-//     <ManageSupportUI
-//       nickNameList={nickNameList}
-//       manageSupportList={manageSupportList}
-//       clearAllChanges={clearAllChanges}
-//       removeAll={removeAll}
-//       handleClose={handleClose}
-//       checked={checked}
-//       setManageSupportList={setManageSupportList}
-//       parentSupportDataList={parentSupportDataList}
-//       getSupportStatusData={getSupportStatusData}
-//       submitNickNameSupportCamps={submitNickNameSupportCamps}
-//       cancelManageRoute={cancelManageRoute}
-//       setSelectedtNickname={setSelectedtNickname}
-//       selectedtNickname={selectedtNickname}
-//       submitButtonDisable={submitButtonDisable}
-//       setUpdatePostion={setUpdatePostion}
-//       unableToFindCamp={unableToFindCamp}
-//     />
-//   );
-
-//   const tagBtn = screen.getByTestId("tag-btn")
-//   expect(tagBtn).toBeInTheDocument()
-
-// });
-it("should render card with title and contents", async () => {
-  const nicknamelist = [
-    {
-      nick_name: "Vikas",
-      id: "1",
-    },
-  ];
-  const manageSupportList = [
-    {
-      topic_num: 1,
-      camp_num: 1,
-      camp_name: "test",
-      support_order: 1,
-      link: "/test",
-    },
-    {
-      topic_num: 2,
-      camp_num: 2,
-      camp_name: "test 2",
-      support_order: 2,
-      link: "/test2",
-    },
-  ];
-  const removeAll = jest.fn();
-  const handleClose = jest.fn();
-  const checked = false;
-  const setManageSupportList = jest.fn();
-  const parentSupportDataList = [
-    {
-      camp_num: 1,
-      support_order: 1,
-      camp_name: "Test",
-      link: "/test",
-    },
-  ];
-  const getSupportStatusData = "test";
-  const submitNickNameSupportCamps = jest.fn();
-  const cancelManageRoute = jest.fn();
-  const setSelectedtNickname = jest.fn();
-  const selectedtNickname = "1";
-  const submitButtonDisable = false;
-  const setUpdatePostion = jest.fn();
-  const unableToFindCamp = false;
-
-  const { getByTestId, getAllByTestId } = render(
-    <ManageSupportUI
-      nickNameList={nicknamelist}
-      manageSupportList={manageSupportList}
-      clearAllChanges={clearAllChanges}
-      removeAll={removeAll}
-      handleClose={handleClose}
-      checked={checked}
-      setManageSupportList={setManageSupportList}
-      parentSupportDataList={parentSupportDataList}
-      getSupportStatusData={getSupportStatusData}
-      submitNickNameSupportCamps={submitNickNameSupportCamps}
-      cancelManageRoute={cancelManageRoute}
-      setSelectedtNickname={setSelectedtNickname}
-      selectedtNickname={selectedtNickname}
-      submitButtonDisable={submitButtonDisable}
-      setUpdatePostion={setUpdatePostion}
-      unableToFindCamp={unableToFindCamp}
-      getManageSupportLoadingIndicator={false}
-    />
-  );
-  const campName = getByTestId("camp_name");
-  expect(campName).toBeInTheDocument();
-
-  const clickOnCamp = getAllByTestId("styles_Bluecolor")[0];
-  expect(clickOnCamp).toBeInTheDocument();
-  fireEvent.click(clickOnCamp);
-  const closeCircle = getAllByTestId("close")[0];
-  expect(closeCircle).toBeInTheDocument();
-  fireEvent.click(closeCircle);
-  // const selectInput = screen.getByRole("combobox");
-
-  //   userEvent.click(selectInput);
-  // const inputE1 = screen.getByTestId("select-option")
-  // const inputE1 = screen.getAllByLabelText("Vikas");
-  // userEvent.click(inputE1[0]);
-
-  // expect(inputE1).toBeInTheDocument();
-  // expect(inputEl).toHaveAttribute("type", "search");
-  // fireEvent.click(inputEl);
-
-  // await fireEvent.change(inputE1, { target: { value: "1" } });
-  //  await  userEvent.selectOptions(getByTestId('select-option'), '1');
-});
 describe("Manage support", () => {
   it("render nick name list", async () => {
     await render(<ManageSupport />);
@@ -657,6 +205,17 @@ describe("Manage support", () => {
       expect(screen.getByText(nickNameList[0].private)).toBeInTheDocument();
     });
   });
+  // it("render handle close function", async () => {
+  //   render(<ManageSupport />);
+  //   await waitFor(() => {
+  //     const clickOnCamp = screen.getAllByTestId("styles_Bluecolor")[0];
+  //     expect(clickOnCamp).toBeInTheDocument();
+  //     fireEvent.click(clickOnCamp);
+  //     const closeCircle = screen.getAllByTestId("close")[0];
+  //     expect(closeCircle).toBeInTheDocument();
+  //     fireEvent.click(closeCircle);
+  //   });
+  // });
   it("render camp record", async () => {
     await render(<ManageSupport />);
     waitFor(async () => {
@@ -675,7 +234,6 @@ describe("Manage support", () => {
       expect(screen.getByText(campRecord.direct_archive)).toBeInTheDocument();
       expect(screen.getByText(campRecord.flag)).toBeInTheDocument();
       expect(screen.getByText(campRecord.go_live_time)).toBeInTheDocument();
-      // expect(screen.getByText(campRecord.is_archive)).toBeInTheDocument();
       expect(screen.getByText(campRecord.is_disabled)).toBeInTheDocument();
       expect(screen.getByText(campRecord.is_one_level)).toBeInTheDocument();
       expect(screen.getByText(campRecord.key_words)).toBeInTheDocument();
@@ -693,28 +251,6 @@ describe("Manage support", () => {
   });
 });
 
-// describe("Test Case", () => {
-//   // beforeEach(()=>{
-//     jest.mock("react-redux", () => ({
-//       ...jest.requireActual("react-redux"),
-//       useSelector: jest.fn().mockImplementation(() => {
-//         return {
-//           campRecord: {
-//             camp_name: "ABC",
-//           },
-//           asof: "",
-//           asofdate: "",
-//           manageSupportUrlLink: "",
-//           currentDelegatedSupportedClick: {},
-//           currentGetCheckSupportExistsData: {
-//             is_delegator: true
-//           },
-//           CurrentCheckSupportStatus: "",
-//           manageSupportStatusCheck: true,
-//         };
-//       }),
-//     }));
-//   // })
 describe("Manage support ui cancle or submit button", () => {
   it("click on cancel button", () => {
     const { getAllByText, container } = render(<ManageSupport></ManageSupport>);
@@ -733,51 +269,24 @@ describe("Manage support ui cancle or submit button", () => {
     ).toBeTruthy();
   });
 
-  it("click on clear all changes button", () => {
-    const { getAllByText, container } = render(<ManageSupport></ManageSupport>);
-    const clear_all_button = getAllByText("Clear all changes")[0];
-    fireEvent.click(clear_all_button);
-    expect(
-      container.getElementsByClassName("ant-select-selection-item")
-    ).toBeTruthy();
-  });
+  // it("click on clear all changes button", () => {
+  //   const { getAllByText, container } = render(<ManageSupport></ManageSupport>);
+  //   const clear_all_button = getAllByText("Clear all changes")[0];
+  //   fireEvent.click(clear_all_button);
+  //   expect(
+  //     container.getElementsByClassName("ant-select-selection-item")
+  //   ).toBeTruthy();
+  // });
 
-  it("click on remove all changes button", () => {
-    const { getAllByTestId, container } = render(
-      <ManageSupport></ManageSupport>
-    );
-    const remove_all_button = getAllByTestId("checkbox")[0];
-    fireEvent.click(remove_all_button);
-    expect(
-      container.getElementsByClassName("ManageSupport_checkbox__DQcrk")
-    ).toBeTruthy();
-  });
+  // it("click on remove all changes button", () => {
+  //   const { getAllByTestId, container } = render(
+  //     <ManageSupport></ManageSupport>
+  //   );
+  //   const remove_all_button = getAllByTestId("checkbox")[0];
+  //   fireEvent.click(remove_all_button);
+  //   expect(
+  //     container.getElementsByClassName("ManageSupport_checkbox__DQcrk")
+  //   ).toBeTruthy();
+  // });
+  
 });
-// describe("Test Cases", () => {
-//   // beforeEach(()=>{
-//     jest.mock("react-redux", () => ({
-//       ...jest.requireActual("react-redux"),
-//       useSelector: jest.fn().mockImplementation(() => {
-//         return {
-//           campRecord: {
-//             camp_name: "ABC",
-//           },
-//           asof: "",
-//           asofdate: "",
-//           manageSupportUrlLink: "",
-//           currentDelegatedSupportedClick: {},
-//           currentGetCheckSupportExistsData: {
-//             is_delegator: false
-//           },
-//           CurrentCheckSupportStatus: "",
-//           manageSupportStatusCheck: true,
-//         };
-//       }),
-//     }));
-//   })
-//   it("click on Submit button - Add or Remove Api", () => {
-//     const { getAllByText } = render(<ManageSupport></ManageSupport>);
-//     const submit_button = getAllByText("Submit")[0];
-//     fireEvent.click(submit_button);
-//   });
-// })
